more changes, making thing clear
